Infotainment-backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
Django==5.0.2
django-channels==0.7.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
oauthlib==3.2.2
obd==0.7.2
Pint==0.20.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pyserial==3.5
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.0
zope.interface==6.1
django-cors-headers==4.3.1
